% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FieldMap.R
\name{FieldMap}
\alias{FieldMap}
\title{Field Map}
\usage{
FieldMap(
  Area,
  site.density,
  site.area,
  overlap = 0.5,
  plot = FALSE,
  areaprecision = 1000
)
}
\arguments{
\item{Area}{vector with horizontal and vertical size \verb{(hor,ver)} of area surveyed in km.}

\item{site.density}{number of sites/km^2. Can be one constant value or vector with two values \verb{(min, max)} to
create a range of densities between simulations.}

\item{site.area}{either:

One values with uniform area for all sites, or

Vector with 4 values \verb{(min, max, mean, st dev)}, to create variable areas. Areas in this case
are normally distributed based on mean and stdev, but within the range of min and max.}

\item{overlap}{proportion of overlap possible between sites: from (0 = no overlap allowed to 1 = sites can occupy same space)}

\item{areaprecision}{value passed to \code{AreaEstimator}. Defines precision of area calculation. Default value (\code{1000}), returns area within 0.1\% of real area occupied by sites}
}
\description{
Creates randomly placed ellipsoid sites in a rectangular field.
}
\details{
\code{FieldMap} creates and plots randomly placed ellipses representing
archaeological sites. The sites are created inside a user-defined rectangle, with random positions
and random rotations. It allows also to control how of overlap between sites.
}
\examples{
#example of map with 8 sites or variable areas and partial overlap
field.example<-FieldMap(Area=c(1,1),site.density=8,site.area=c(50000,250000,150000,50000),overlap=0.5,plot=TRUE)

}
