% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametersCreator.R
\name{ParametersCreator}
\alias{ParametersCreator}
\title{Parameters creator}
\usage{
ParametersCreator()
}
\arguments{
\item{col.width}{the space between columns in the grid IN METERS}

\item{grid.type}{options are: "square","rectangle","staggered","hexagonal","arbitrary.staggered", following Kintigh 1988}

\item{simulations}{number of random maps to be created and contrasted with the grids}

\item{Area}{vector with horizontal and vertical size of area surveyed in km
OBS: Sites will all be ellipses with radii not too different and random angles}

\item{SiteDensity}{measured as number of sites/km^2. Can be either one value or a vector with 2 values (min and max) to create a range of densities}

\item{site.area}{can be one of two options: 1. one value indicating the area of all sites, in meter^2;
or 2. a vector with 4 values: min, max, mean (or median), and standard deviation in meter^2}

\item{overlap}{maximum overlap of site area, ranging from 0 = no overlap allowed to 1 = complete overlap possible}

\item{obj.density}{artifacts per m^2. Can be a single value (uniform for all sites) or a range of values defined as min and max}

\item{obj.distribution}{type of cloud distribution for artifacts inside sites. Choose from: 'uniform', 'linear', 'spherical', 'sinusoidal'}

\item{survey.radius}{the radius of the survey pit (assumed to be a circle)}
}
\description{
Creates the List of Parameters for SurveySim
}
\details{
This function will create a user text interface to define all the arguments required
by the SurveySim function. It generates a list of class \code{SurveySim()}.
All parameters are loaded inside the function.
}
\examples{
ParametersCreator()
}
\references{
Kintigh (1988) The Effectiveness of Subsurface Testing: A Simulation Approach.
American Antiquity, 53:686-707.
}
