% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametersExample.R
\docType{data}
\name{parametersExample}
\alias{parametersExample}
\title{Parameters Example}
\format{
An object of class \code{surveySim} of length 10.
}
\usage{
parametersExample
}
\description{
A test list of parameters for \code{surveySim()}
}
\details{
This is just a test list of parameters to be used as an example in \code{surveySim()}.

The values of this parameters are:

Width between survey lines (\code{col.width}) = \strong{50 m}

Type of survey grid (\code{grid.type}) = \strong{hexagonal}

Number of \code{simulations} = \strong{10}

Survey \code{area} = \strong{0.5km x 0.5km}

Density of sites (\code{site.density}) = \strong{20}

Area of sites (\code{site.area}) = \strong{10,000\ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}}}

Maximum site \code{overlap} = \strong{0.5}

Density of artifacts (\code{obj.density}) = \strong{1/\ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}}}

Artifact distribution (\code{obj.distribution}) = \strong{spherical}

Survey radius (\code{survey.radius}) = \strong{0.5m}
}
\keyword{datasets}
