% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummaryPlot.R
\name{PlotSurveySumm}
\alias{PlotSurveySumm}
\title{Plot Survey Summaries}
\usage{
PlotSurveySumm(SummaryList, plot = "sites.found", labels = NULL)
}
\arguments{
\item{SummaryList}{a list of survey summaries, the output of \code{SurveySim()}}

\item{plot}{what variable to plot. Options are:

\code{sites.found} - plots frequency of sites found

\code{survey.hits} - plots frequency of surveys that hit a site

\code{success.rate.index} - plots the success rate of surveys, i.e. the ratio of successful surveys over total surveys

\code{sites.found.ARTI} - plots frequency of sites found \strong{based on artifact present in sites}

\code{survey.hits.ARTI} - plots frequency of surveys that found \strong{at least one artifact}}

\item{labels}{vector with name of each item in list, to be added to the legend. If \code{NULL}, names will be taken from list}
}
\description{
Plots the different results from \code{SurveySim()}
}
\details{
This function will plot the results of the \code{SurveySim} Simulations using Kernel Density plots.
All the grids that are to be compared should be grouped into 1 list (list(a,b,c,etc...))
The plot function allows you to choose different parts of the survey summaries produced by \code{SurveySim()} you want to plot.
}
\examples{
 #create 3 Simulations with sites of different sizes:
 small.sites<-ParametersExample
 small.sites$site.area=500

 medium.sites<-ParametersExample
 medium.sites$site.area=1000

 large.sites<-ParametersExample
 large.sites$site.area=2000

 #run the 3 simulations
 small.survey<-SurveySim(small.sites)
 medium.survey<-SurveySim(medium.sites)
 large.survey<-SurveySim(large.sites)

 #create the comparative plot.
 #note that the results go into a list. If labels are not given, legend is built on list names
 PlotSurveySumm(list(small.survey,medium.survey,large.survey),plot="sites.found",labels=c("Small sites","Medium sites","Large sites"))

}
