% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveySim.R
\name{surveySim}
\alias{surveySim}
\title{Survey Simulations}
\usage{
surveySim(
  survey.parameters,
  artifact.analysis = TRUE,
  plot = TRUE,
  plot.artifacts = FALSE,
  areaprecision = 1000
)
}
\arguments{
\item{survey.parameters}{List of class \code{surveySim} generated by \code{ParametersCreator()}}

\item{artifact.analysis}{Includes the analysis of artifacts in sites. Can be very time-consuming for mid-high artifact densities. Default = \code{TRUE}}

\item{plot}{If \code{TRUE} the last iteration of the simulations will be plotted.}

\item{plot.artifacts}{If \code{TRUE} the artifacts in each site will be plotted. Default = \code{FALSE} because it is resource intensive for mid-high densities.}

\item{areaprecision}{Area covered by sites is approximated by a cookie-cutter approach. Default precision = \code{1000}
gives approximate area within 1\% error. Higher values reduce error but are more resource intensive.}
}
\value{
A list with four objects: \tabular{ll}{
\code{Summary} \tab A matrix with summary statistics about number of surveys, frequency of site founds, artifacts presence, and success rate of simulations.\cr
\tab \cr
\code{BySite} \tab A matrix with results of the analyses by site from each of the survey areas created in the simulation. \cr
\tab \cr
\code{ByArtifact} \tab A matrix with results of the analyses by artifact presence in survey pits from each of the survey areas created in the simulation. \cr
\tab \cr
\code{Parameters} \tab A list with the parameters used to run the simulation. \cr
}
}
\description{
Master function that runs survey simulations
}
\details{
\code{surveySim} is the main function that will conduct the simulations
of survey grids based on the Survey Parameters created by user.
}
\examples{
#Runs simulations based on parametersExample
surveyExample<-surveySim(parametersExample)

#Edit parametersExample to have 50 simulations and then run simulations
tmp_parameters <-parametersExample
tmp_parameters$simulations<-50
survey50<-surveySim(tmp_parameters)

}
