% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametersCreator.R
\name{parametersCreator}
\alias{parametersCreator}
\title{Parameters creator}
\usage{
parametersCreator()
}
\value{
A list of class \code{SurveySim} with simulation parameters.
}
\description{
Creates the List of Parameters for \code{surveySim()}
}
\details{
This function will create a user text interface to define all the arguments required
by the \code{surveySim} function. It generates a list of class \code{surveySim()}.
All parameters of this list are loaded inside the function.
\itemize{
\item \code{col.width} the space between columns in the grid IN METERS
\item \code{grid.type} options are: "square","rectangle","staggered","hexagonal","arbitrary.staggered", following Kintigh 1988
\item \code{simulations} number of random maps to be created and contrasted with the grids
\item \code{area} vector with horizontal and vertical size of area surveyed in km
OBS: Sites will all be ellipses with radii not too different and random angles
\item \code{site.density} measured as number of sites/\ifelse{html}{\out{km<sup>2</sup>}}{\eqn{km^2}}. Can be either one value or a vector with 2 values (min and max) to create a range of densities
\item \code{site.area} can be one of two options: 1. one value indicating the area of all sites, in \ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}};
or 2. a vector with 4 values: min, max, mean (or median), and standard deviation in \ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}}
\item \code{overlap} maximum overlap of site area, ranging from 0 = no overlap allowed to 1 = complete overlap possible
\item \code{obj.density} artifacts per \ifelse{html}{\out{m<sup>2</sup>}}{\eqn{m^2}}. Can be a single value (uniform for all sites) or a range of values defined as min and max
\item \code{obj.distribution} type of cloud distribution for artifacts inside sites. Choose from: 'uniform', 'linear', 'spherical', 'sinusoidal'
\item \code{survey.radius} the radius of the survey pit (assumed to be a circle)
}
}
\examples{
parametersCreator()
}
\references{
Kintigh (1988) The Effectiveness of Subsurface Testing: A Simulation Approach.
American Antiquity, 53:686-707.
}
