% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParametersExample.R
\docType{data}
\name{ParametersExample}
\alias{ParametersExample}
\title{Parameters Example}
\format{
An object of class \code{SurveySim} of length 10.
}
\usage{
ParametersExample
}
\description{
A test of parameters for \code{SurveySim()}
}
\details{
This is just a test list of parameters to be used as an example in \code{SurveySim()}.

The values of this parameters are:

Width between survey lines (\code{col.width}) = \strong{50 m}

Type of survey grid (\code{grid.type}) = \strong{hexagonal}

Number of \code{simulations} = \strong{25}

Survey \code{area} = \strong{0.5km x 0.5km}

Density of sites (\code{site.density}) = \strong{20}

Area of sites (\code{site.area}) = \strong{10,000 m^2}

Maximum site \code{overlap} = \strong{0.5}

Density of artifacts (\code{obj.density}) = \strong{1/m^2}

Artifact distribution (\code{obj.distribution}) = \strong{spherical}

Survey radius (\code{survey.radius}) = \strong{0.5m}
}
\keyword{datasets}
