% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurveySim.R
\name{SurveySim}
\alias{SurveySim}
\title{Survey Simulations}
\usage{
SurveySim(
  SurveyParameters,
  artifact.analysis = TRUE,
  plot = TRUE,
  plot.artifacts = FALSE,
  areaprecision = 1000
)
}
\arguments{
\item{SurveyParameters}{List of class \code{SurveySim} generated by \code{SurveyParameters()}}

\item{artifact.analysis}{Includes the analysis of artifacts in sites. Can be very time-consuming for mid-high artifact densities. Default = \code{TRUE}}

\item{plot}{If \code{TRUE} the last iteration of the simulations will be plotted.}

\item{plot.artifacts}{If \code{TRUE} the artifacts in each site will be plotted. Default = \code{FALSE} because it is resource intensive for mid-high densities.}

\item{areaprecision}{Area covered by sites is approximated by a cookie-cutter approach. Default precision = \code{1000}
gives approximate area within 1\% error. Higher values reduce error but are more resource intensive.}
}
\description{
Master function that runs survey simulations
}
\details{
\code{SurveySim} is the main function that will conduct the simulations
of survey grids based on the Survey Parameters created by user.
}
\examples{
#Runs simulations based on `ParametersExample`
SurveySim(ParametersExample)

#Edit ParametersExample to have 50 simulations and then run simulations
tmp_parameters <-ParametersExample
tmp_parameters$simulations<-50
SurveySim(tmp_parameters)

}
